create output_parser.rs
