+++ lib/judge_core/src/logic.rs
@@ -6,3 +6,4 @@ pub mod registered_procedure_converter;
 pub mod registered_procedure_remover;
 pub mod runner;
 pub mod writer_schema_registerer;
+pub mod output_parser;
